apps/llm-service/requirements.txt
redis>=5.0.0
psycopg[binary]>=3.1.0
openai>=1.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
